TestesUnitarios.Desafio.Console/Services/ValidacoesString.cs
namespace TestesUnitarios.Desafio.Console.Services
{
    public class ValidacoesString
    {
        public int RetornarQuantidadeCaracteres(string texto)
        {
            var numeroCaracteres = texto.Length; // método trata do tamanho da string
            return numeroCaracteres;
        }

        public bool ContemCaractere(string texto, string textoProcurado)
        {
            var contem = texto.Contains(textoProcurado); // método diz que uma string maior CONTÉM uma menor
            return contem;
        }

        public bool TextoTerminaCom(string texto, string textoProcurado)
        {
            var termina = texto.EndsWith(textoProcurado); // método diz que uma string maior TERMINA COM uma menor
            return termina;
        }
    }
}